feat(auth-route): rename
