+++ examples/shop-with-blog/client/src/components/AuthorizedRoute.js
